requirements.txt
python-binance==1.0.19
python-dotenv==1.0.0
rich==13.7.0
pydantic==2.6.1
pydantic-settings==2.2.1 